move app/*/project/edit => app/*/edit and genericize accordingly
